backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
beanie>=1.25.0
httpx>=0.27.0
pytest>=8.0.0
